Create foodOrderingSystem.c
This is a mock example of how I would program a food ordering software.
The program displays their menu and lets the user order food.
For certain food items, the user can make combo meals for an additional cost.
The program calculates the total of all food items ordered and adds a 6% sales tax.